Create tools and move clean function in it
