+++ games/FlagGuesser/FlagGuesser.py
@@ -1,11 +1,11 @@
 from games.Game import Game
 from domain.Constants import LANGUAGE, DIFFICULTY, CONTINENT
+from domain.Tools import clean
 from infra.FlagRepository import FlagRepository
 
 from discord import Thread, File, DMChannel
 from discord.ext.commands import Context
 
-from unidecode import unidecode
 import io
 import aiohttp
 import random
@@ -100,14 +100,3 @@ For this game, you have to give the name of the country in **{LANGUAGE.trad(self
 	async def end(self):
 		self.has_ended = True
 		await self.send_sum_up()
-
-
-def clean(s: str) -> str:
-	s = unidecode(s) 		# transform all accented letters to non-accented letters
-	s = s.lower()			# we want the string to be lowercase
-	s = s.replace("-", " ") # change hyphens to spaces
-	s = str.rstrip(s) 		# remove trailing whitespace characters
-	s = str.lstrip(s) 		# remove leading whitespace characters
-	# remove all multiple spaces following each other
-	s = "".join([s[i] for i in range(len(s)) if i==0 or s[i].isalpha() or (s[i]==" " and s[i-1]!=" ")])
-	return s
